2025/boot_imx93.sh: modify PATH to be able to use local uuu
The uuu tool is necessary for running this script, and while it is
provided locally, the script was unable to find it and was throwing "uuu
command not found". This commit fixes that by modifying PATH for the
duration of the script. The alternative would have been to use "./uuu",
however this approach seems cleaner to me.

Signed-off-by: Panait Teodor Dimitrie <teodor.panait@stud.acs.upb.ro>